pyproject.toml
[tool.poetry]
name = "multi_agent_system_demo"
version = "0.1.0"
description = "Multi Agent System Demo"
authors = ["Avinash Vaidya"]
readme = "README.md"
packages = [{ include = "mas_autogen" }]

[tool.poetry.dependencies]
python = "~3.12"
fastapi = "^0.103.0"
uvicorn = "^0.23.0"
requests = "^2.31.0" 
autogen-agentchat = "~0.2"  # Microsoft AutoGen (Multi-Agent Chat System)
python-dotenv = "^1.0.0"
loguru = "^0.7.0"  
pydantic = "^2.4.2" 

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"
mypy = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"